src/main/webapp/app/entities/shopping-cart/delete/shopping-cart-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { ShoppingCartService } from '../service/shopping-cart.service';

import { ShoppingCartDeleteDialogComponent } from './shopping-cart-delete-dialog.component';

describe('Component Tests', () => {
  describe('ShoppingCart Management Delete Component', () => {
    let comp: ShoppingCartDeleteDialogComponent;
    let fixture: ComponentFixture<ShoppingCartDeleteDialogComponent>;
    let service: ShoppingCartService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [ShoppingCartDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(ShoppingCartDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(ShoppingCartDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(ShoppingCartService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          spyOn(service, 'delete').and.returnValue(of({}));

          // WHEN
          comp.confirmDelete(123);
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith(123);
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
